fix: move `WasmBytes` to fluvio-protocol crate
