+++ crates/fluvio-protocol/src/core/mod.rs
@@ -1,12 +1,15 @@
 mod decoder;
 mod encoder;
 mod varint;
+mod wasm_bytes;
 mod zerocopy;
 
 pub use self::decoder::Decoder;
 pub use self::decoder::DecoderVarInt;
 pub use self::encoder::Encoder;
 pub use self::encoder::EncoderVarInt;
+pub use self::wasm_bytes::WasmBytes;
+
 pub type Version = i16;
 
 #[cfg(test)]
