+++ crates/fluvio-protocol/src/lib.rs
@@ -33,5 +33,6 @@ pub use self::core::DecoderVarInt;
 pub use self::core::Encoder;
 pub use self::core::EncoderVarInt;
 pub use self::core::Version;
+pub use self::core::WasmBytes;
 
 pub use bytes;
